# state/cache: implement get_many/set_many