--- a/src/dvc_data/hashfile/cache.py
+++ b/src/dvc_data/hashfile/cache.py
@@ -1,7 +1,9 @@
 import os
 import pickle
+from collections.abc import Iterable, Iterator
 from functools import wraps
-from typing import Any, Optional
+from itertools import zip_longest
+from typing import Any, ClassVar, Literal, Optional
 
 import diskcache
 from diskcache import Disk as _Disk
@@ -9,6 +11,8 @@ from diskcache import (
     Index,  # noqa: F401
     Timeout,  # noqa: F401
 )
+
+from dvc_data.compat import batched
 
 
 class DiskError(Exception):
@@ -56,8 +60,62 @@ class Cache(diskcache.Cache):
         **settings: Any,
     ) -> None:
         settings.setdefault("disk_pickle_protocol", 4)
+        settings.setdefault("cull_limit", 0)
         super().__init__(directory=directory, timeout=timeout, disk=disk, **settings)
         self.disk._type = self._type = type or os.path.basename(self.directory)
 
     def __getstate__(self):
         return (*super().__getstate__(), self._type)
+
+
+class HashesCache(Cache):
+    SQLITE_MAX_VARIABLE_NUMBER: ClassVar[Literal[999]] = 999
+    """The maximum number of host parameters is 999 for SQLite versions prior to 3.32.0
+    (2020-05-22) or 32766 for SQLite versions after 3.32.0.
+
+    Increasing this number does not yield any performance improvement, so we leave it at
+    the old default.
+    """
+
+    def get_many(
+        self, keys: Iterable[str], default=None
+    ) -> Iterator[tuple[str, Optional[str]]]:
+        if self.is_empty():
+            yield from zip_longest(keys, [])
+            return
+
+        for chunk in batched(keys, self.SQLITE_MAX_VARIABLE_NUMBER):
+            params = ", ".join("?" * len(chunk))
+            query = f"SELECT key, value FROM Cache WHERE key IN ({params}) and raw = 1"  # noqa: S608
+            d = dict(self._sql(query, chunk).fetchall())
+            for key in chunk:
+                yield key, d.get(key, default)
+
+    def set_many(self, items: Iterable[tuple[str, str]], retry: bool = False) -> None:
+        with self.transact(retry):
+            self._con.executemany(
+                "INSERT OR REPLACE INTO Cache("
+                " key, raw, store_time, expire_time, access_time,"
+                " tag, mode, filename, value"
+                ") VALUES (?, 1, 0, null, 0, null, 1, null, ?)",
+                items,
+            )
+
+    def is_empty(self) -> bool:
+        res = self._sql("SELECT EXISTS (SELECT 1 FROM Cache)")
+        ((exists,),) = res
+        return exists == 0
+
+    def set(
+        self, key: str, value: str, expire=None, read=False, tag=None, retry=False
+    ) -> Literal[True]:
+        self.set_many([(key, value)], retry=retry)
+        return True
+
+    def get(
+        self, key, default=None, read=False, expire_time=False, tag=False, retry=False
+    ):
+        cursor = self._sql("SELECT value FROM Cache WHERE key = ? and raw = 1", (key,))
+        if rows := cursor.fetchall():
+            return rows[0][0]
+        return default