--- a/src/dvc_data/hashfile/db/__init__.py
+++ b/src/dvc_data/hashfile/db/__init__.py
@@ -16,6 +16,7 @@ if TYPE_CHECKING:
     from dvc_objects.fs.base import AnyFSPath, FileSystem
     from fsspec import Callback
 
+    from dvc_data.hashfile.state import StateBase
     from dvc_data.hashfile.tree import Tree
 
     from .index import ObjectDBIndexBase
@@ -56,7 +57,7 @@ class HashFileDB(ObjectDB):
         from dvc_data.hashfile.state import StateNoop
 
         super().__init__(fs, path, read_only=read_only)
-        self.state = config.get("state", StateNoop())
+        self.state: StateBase = config.get("state", StateNoop())
         self.verify = config.get("verify", self.DEFAULT_VERIFY)
         self.cache_types = config.get("type") or copy(self.DEFAULT_CACHE_TYPES)
         self.slow_link_warning = config.get("slow_link_warning", True)
@@ -107,19 +108,22 @@ class HashFileDB(ObjectDB):
             **kwargs,
         )
 
-        for o in oids:
-            cache_path = self.oid_to_path(o)
+        oid_cache_paths = {o: self.oid_to_path(o) for o in oids}
+        for o, cache_path in oid_cache_paths.items():
             try:
                 if verify:
                     self.check(o, check_hash=True)
                 self.protect(cache_path)
-                self.state.save(
-                    cache_path,
-                    self.fs,
-                    HashInfo(name=self.hash_name, value=o),
-                )
             except (ObjectFormatError, FileNotFoundError):
                 pass
+
+        self.state.save_many(
+            (
+                (cache_path, HashInfo(name=self.hash_name, value=o), None)
+                for o, cache_path in oid_cache_paths.items()
+            ),
+            self.fs,
+        )
         return transferred
 
     def protect(self, path):