This PR implements batching `save` and `get` in `State`.

As a result of this change, `_build_tree()` is faster by 70% - 90% on the first run, and 50% faster on repeated runs with this PR. You can find the results below for the mnist dataset. 

This should dramatically improve the performance for `dvc add --no-relink`, and other places where we build trees and compute hashes. Real-world impact is mixed and depends on individual commands, but all of the commands show improvements from 4-5 % to 20%, and in some cases,  even 70% or more.

This also brings back support for `core.checksum_jobs` to concurrently hash multiple files at once.

This is done by partitioning the files to hash into two buckets: a) files that are smaller than 1MB and, b) files that are larger than 1MB. 

Files smaller than 1MB are hashed sequentially, and larger files are hashed in a ThreadPoolExecutor.

### Benchmark

### `main`

```
$ rm -rf .dvc/tmp
$ time dvc-data build dataset
object e42412b82dcab425ce9c7e2d0abdfb78.dir

________________________________________________________
Executed in   16.39 secs    fish           external
   usr time   11.18 secs    0.00 micros   11.18 secs
   sys time    3.73 secs  348.00 micros    3.73 secs

$ time dvc-data build dataset
object e42412b82dcab425ce9c7e2d0abdfb78.dir

________________________________________________________
Executed in    2.76 secs    fish           external
   usr time    2.49 secs  236.00 micros    2.49 secs
   sys time    0.22 secs   52.00 micros    0.22 secs
```

### `PR`

```
$ rm -rf .dvc/tmp
$ time dvc-data build dataset
object e42412b82dcab425ce9c7e2d0abdfb78.dir

________________________________________________________
Executed in    2.95 secs    fish           external
   usr time    2.33 secs  262.00 micros    2.33 secs
   sys time    0.36 secs   61.00 micros    0.36 secs

$  time dvc-data build dataset
object e42412b82dcab425ce9c7e2d0abdfb78.dir

________________________________________________________
Executed in    1.36 secs    fish           external
   usr time    1.20 secs  198.00 micros    1.20 secs
   sys time    0.12 secs   38.00 micros    0.12 secs
```


<details><summary>DVC-bench results on an MNIST dataset</summary>
<p>

| Test | main (in s) | PR (in s) | △ |
|--------|--------|--------|--------|
| test_add-add | 87.2241 | 62.7932 | 28% |
| test_checkout_copy-checkout | 31.0993 | 29.1070 | 6% |
| test_checkout_copy-checkout-noop | 6.5580 | 5.4686 | 16% |
| test_checkout_copy-checkout-update | 7.5199 | 6.2810 | 16% |
| test_data_status-data-new | 21.3120 | 15.9083 | 25% | 
| test_data_status-data-noop | 5.4173 | 4.1119 | 24% | 
| test_data_status-data-changed | 5.3719 | 4.1306 | 23% | 
| test_data_status-data-changed-noop | 5.3541 | 4.1927 | 21% | 
| test_diff-diff | 22.6914 | 18.3842 | 19% | 
| test_diff-diff-changed | 6.5643 | 5.3981 | 17% | 
| test_diff-diff-changed-noop | 6.5565 | 5.3985 | 17% | 
| test_diff-diff-noop | 6.6704 | 5.3392 | 20% | 
| test_import_url-import-url | 131.6878 | 87.0350 | 34% | 
| test_status-status | 20.6743 | 6.0155 | 71% | 
| test_status-status-changed | 6.0105 | 3.1429 | 48% | 
| test_update-import-url | 130.6300 | 86.5345 | 33% | 
| test_fetch-fetch | 37.0039 | 26.7890 | 27% |


See https://github.com/iterative/dvc-bench/actions/runs/10252645782.
</p>
</details>