@skshetry 2024-07-30
> what is the context for this?

At the moment, we save each state entries one at a time, and each individual save is within their own transaction. If you have a lot of files, this can be as costly as doing copy/symlink/reflink. (On my machine, it takes 1872 us for a `__setitem__`, 2103 us for a `reflink`).

(See https://github.com/iterative/dvc/issues/10491, https://github.com/iterative/dvc/issues/6227 and tens of other issues).

The idea is to try to save all the entries at once, at least for all files within a directory at a certain level. (We already assume that the dataset directory that user's are tracking are shallow and each level has tons of files. This goes with what we see in real world with sharding datasets into different directories like `train/0/`, `train/1` etc.).

But to take advantage of this, we need to also change how we build objects for a tree (which happens in `build_tree`). Instead of doing everything inside a for-loop, we need to separate the operation into multiple steps, with all of them doing batching that includes:

1. getting all saved hashes from State
2. figuring out files to hash (if they are missing from State).
3. hashing all files at once
4. saving all hashes back to State.
5. building the tree object.

"hashing all files at once" will also gives us ability to concurrently hash files. We can even split this into two stages: hash small files (< 1MB) sequentially, and hash larger files in a `ThreadPoolExecutor`. (Note that we already have to gather `fs.info()` of all the files while building the object which already has `size` information).

> what is the status of this?

That's a matter of priority that we set. If it gets prioritized, I can work on this, initially focusing on this and improving performance for `dvc add`.