@dberenbaum 2024-07-30
> which happens in `build_tree`

Do you mean `build_entry` (https://github.com/iterative/dvc-data/blob/ffa6839e35cdd193da469605978eb8e0946433ee/src/dvc_data/index/build.py#L19)?