@skshetry 2024-07-30
> I don't see that much time is spent in `build_tree`.


Both do similar work, but one builds the `tree` object, while other builds the index.

`build_tree` is for non-index based operations. That is used in `dvc add`, `dvc commit` and `dvc repro`. In `dvc repro`, it's even worse because we rebuild objects many times.

`build_entries` is for index-based operations which is used by `data status`, `dvc checkout`, etc.