@dberenbaum 2024-07-30
Is it `hashfile.build._build_tree` then? I saw `index.save.build_tree` initially, but that doesn't make sense to me.