@shcheklein 2024-07-31
@skshetry yes, I was looking into the disckcache and build_tree recently. It all makes sense to me if it's true that disckcache can't handle many parallel operations (may be with some SQLite changes). If there is not obvious way to check this (I was about to change `dvc-data` to add TQDM first and make it parallel to play a bit with SQLite versions / settings) - it's fine to go this path and do batching (+ also making it parallel in terms of hashes).

DiskCache is really advertising multithreading, really good performance, etc - so it's surprising to see that it's just getting stuck on a "basic" parallel inserts so much. If this is about transactions - can we check if we can run everything in a single transaction then (and if it fails - start over? or do partial commits if possible each 1000?)

Are there some SQLite settings we can tune / disable that we don't need since we are not a web app?

Does DisckCahe create any indecies? Can you share the structure of the DB underneath?