@skshetry 2024-07-31
DiskCache is fast for web applications like django apps, etc. as a replacement for redis. 
On those web applications, you have a very different workload, where you are unlikely to save millions of rows, or read 10s of millions of rows at once.

Using sqlite in parallel does not work that well. There is python thread overhead, and on read, returning rows has to be converted to python objects which is going to be blocked by GIL. On write, it is going to be serialized on sqlite3 side.

>  If this is about transactions - can we check if we can run everything in a single transaction then (and if it fails - start over? or do partial commits if possible each 1000?)

Yes, using transactions help (reduces as much by 10x), but you will still have to call `Connection.execute` a million times. We can do better by using `executemany`.


> Are there some SQLite settings we can tune / disable that we don't need since we are not a web app?

I tried changing some settings but nothing helped when I worked on dvc-data-rs.

> Does DisckCahe create any indecies? Can you share the structure of the DB underneath?

Yes, it does. I find it useful for our read-heavy workload. 

You can see the structure here: https://github.com/skshetry/dvc-data-rs/blob/4be70284f7c7be1a11e0d8c250a7ea3df4cdeb42/src/state.rs#L43-L66.