@skshetry 2024-07-31
> Is it `hashfile.build._build_tree` then? I saw `index.save.build_tree` initially, but that doesn't make sense to me.

Sorry, yes it is `_build_tree`.

https://github.com/iterative/dvc-data/blob/ffa6839e35cdd193da469605978eb8e0946433ee/src/dvc_data/hashfile/build.py#L90