@shcheklein 2024-07-31
> Yes, using transactions help (reduces as much by 10x), but you will still have to call Connection.execute a million times. We can do better by using executemany.

no doubt about that (though we might loose a bit since it will be serials - first read, then md5, then write) vs all in parallel.

just to be mindful about the complexity of changes if we can do 10-15x with simple changes - I would do that first. That's a huge difference and can unblock many people and we can do it "quick".