@skshetry 2024-08-01
> (though we might loose a bit since it will be serials - first read, then md5, then write) vs all in parallel.

Parallelizing a large function that does a lot of things, and has different characteristics (IO-bounded + CPU-bounded) does not work well as we have found out in dvc. It has tons of overheads, hides a lot of things, makes it difficult to optimize and most importantly, does not scale linearly. Also, not everything is parallelizable as we see with sqlite3. 

If you look into our internals, we have slowly moved away from doing that, and instead, we have added batching support to critical APIs like `fs.info()`, `transfer`, etc.
Our plan was/is to optimize those APIs if needed (maybe by dropping to C extensions or looking for a different solutions or maybe using asyncio/threads in those individual functions).

---

Regarding "loose a bit" thing, if it's going to be faster than running it in parallel, then we won't lose anything. Running some computations to check what files to hash are going to be cheaper than going to sqlite3, or running on another thread with GIL.