@skshetry 2024-08-06
> Thanks @skshetry! Do you have any examples or benchmarks to show the overall improvement and in what scenarios it should be better? 

See the description above (at the very end) and https://github.com/iterative/dvc-bench/actions/runs/10252645782.

> Are there any scenarios where you expect it to be worse?

`INSERT OR REPLACE INTO` is usually slower than what diskcache does if there are existing `keys`. Diskcache first `select`s key and either `insert`s or `update`s the row.
This is likely due to unique index constraints that we have on `(key, raw)`.

And now `_build_file`/`_build_tree` goes through more code and stages, so in small microbehchmarks, they may be slightly slower. But this is unlikely to show up in real-world.
(Check out the above benchmarks).